+++ Main App/ViewControllers/TimerSetupController.swift
@@ -97,6 +97,7 @@ class TimerSetupController: TimerSetPickerController {
         
         self.alarmSetupButton.isMusicOn = .Music == self.timerObject.soundMode
         self.alarmSetupButton.isSoundOn = .Sound == self.timerObject.soundMode
+        self.alarmSetupButton.isTicksOn = self.timerObject.audibleTicks
         self.alarmSetupButton.isVibrateOn = (.VibrateOnly == self.timerObject.alertMode || .Both == self.timerObject.alertMode)
         self.colorThemePicker.reloadAllComponents()
         self.warningThresholdTimePicker.reloadAllComponents()