+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - I removed the stop timer when going in the background, so that means the timer pauses in the background.
 - The window no longer goes to the timer list when the app backgrounds (but the timer will pause).
 - The splash screen looks better.
+- Added an image for audible ticks in the sound mode button.
 
 ***Version 1.0.0.2020* ** *(December 19, 2018)*
 - More work on setting up accessibility.